add setting of flag in bulk scan case creation plus refactor of class name
