+++ src/test/java/uk/gov/hmcts/probate/controller/BusinessValidationUnitTest.java
@@ -26,7 +26,7 @@ import uk.gov.hmcts.probate.service.CaseEscalatedService;
 import uk.gov.hmcts.probate.service.CaseStoppedService;
 import uk.gov.hmcts.probate.service.ConfirmationResponseService;
 import uk.gov.hmcts.probate.service.EventValidationService;
-import uk.gov.hmcts.probate.service.HandOffLegacyService;
+import uk.gov.hmcts.probate.transformer.HandOffLegacyTransformer;
 import uk.gov.hmcts.probate.service.NotificationService;
 import uk.gov.hmcts.probate.service.StateChangeService;
 import uk.gov.hmcts.probate.service.caseaccess.AssignCaseAccessService;
@@ -146,7 +146,7 @@ class BusinessValidationUnitTest {
     @Mock
     private AssignCaseAccessService assignCaseAccessService;
     @Mock
-    private HandOffLegacyService handOffLegacyService;
+    private HandOffLegacyTransformer handOffLegacyTransformer;
 
     private BusinessValidationController underTest;
 
@@ -179,7 +179,7 @@ class BusinessValidationUnitTest {
             ihtEstateValidationRule,
             solicitorPostcodeValidationRule,
             assignCaseAccessService,
-            handOffLegacyService);
+            handOffLegacyTransformer);
 
         when(httpServletRequest.getRequestURI()).thenReturn("/test-uri");
     }
